requirements.txt
asgiref==3.9.1
beautifulsoup4==4.13.4
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
django-environ==0.12.0
djangorestframework==3.16.1
google_search_results==2.4.2
greenlet==3.2.4
idna==3.10
playwright==1.54.0
playwright-stealth==2.0.0
psycopg2-binary==2.9.10
pyee==13.0.0
python-dotenv==1.1.1
requests==2.32.4
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
